tools/voice_backend/pyproject.toml
[project]
name = "copilot-voice-backend"
version = "1.0.0"
description = "WebRTC voice backend for Copilot ESP32-S3"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiortc>=1.6.0",
    "aiohttp>=3.9.0",
    "websockets>=14.0",
    "av>=12.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
voice-backend = "src.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
